estate-src/requirements.txt
Django==4.2 
djangorestframework==3.14.0
django-environ==0.9.0
django-filter==23.1
django-autoslug==1.9.8
django-countries==7.2.1 
Pillow==9.5.0
django-phonenumber-field==6.0.0
phonenumbers==8.13.0
flake8==6.0.0
black==24.10.0
isort==5.12.0
djoser==2.1.0 
djangorestframework-simplejwt==5.0.0
PyJWT==2.6.0
pytest-django==4.5.0
pytest-factoryboy==2.1.0 
Faker==18.0.0
pytest-cov==4.0.0
redis==5.0.0
# celery==5.3.1
# flower==1.1.2 
django-celery-email==3.0.0
argh==0.26.2 
PyYAML==6.0
watchdog==2.1.9